JAVA: Renamed and moved the MarkupEncoderUtil to the appropriate place.
