+++ alert-common/src/main/java/com/synopsys/integration/alert/common/util/MarkupEncoderUtil.java
@@ -53,4 +53,5 @@ public class MarkupEncoderUtil {
 
         return newTxt.toString();
     }
+
 }
